packages/ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.0
langchain==0.1.0
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.1.0
redis==5.0.1
httpx==0.25.1
sentry-sdk[fastapi]==1.39.1
prometheus-client==0.19.0
psutil==5.9.6
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
